requirements.txt
Flask==1.1.2
requests==2.23.0
boto3==1.13.6
ecdsa==0.15
PyJWT[crypto]==1.7.1
